requirements.txt
pynput>=1.7.6
screeninfo>=0.8.1
pyperclip>=1.8.2
PyYAML>=6.0
colorlog>=6.7.0
pywin32>=305; sys_platform == 'win32'
